% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classdefns.R
\docType{methods}
\name{as.data.frame,mortalityFits-method}
\alias{as.data.frame,mortalityFits-method}
\title{function for as.data.frame applied to a mortalityFits
object}
\usage{
\S4method{as.data.frame}{mortalityFits}(x, row.names = NULL,
  optional = FALSE, ...)
}
\description{
function for as.data.frame applied to a mortalityFits
object
}

