% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mortalitymodels.R
\docType{data}
\name{binomial.models}
\alias{binomial.models}
\title{this is a list, provided for convenience, of all of
the models based on the binomial likelihood that are
automatically created}
\format{list}
\usage{
binomial.models
}
\description{
this is a list, provided for convenience, of all of
the models based on the binomial likelihood that are
automatically created
}
\keyword{datasets}

