% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fitmethod-combined.R
\name{combined.fit}
\alias{combined.fit}
\title{combined.fit}
\usage{
combined.fit(model.obj, data, M = 5, delta = 0.1, grid.dim = 15,
  verbose = TRUE, ignore.folded = FALSE, ...)
}
\arguments{
\item{model.obj}{a mortalityModel object}

\item{data}{a mortalityData object}

\item{M}{TODO}

\item{delta}{TODO}

\item{grid.dim}{TODO}

\item{verbose}{TODO}

\item{ignore.folded}{TODO}

\item{...}{other params to pass to optimFit and gridSearchFit's
fit methods}
}
\value{
a mortalityFitOptim object
}
\description{
use a combination of optim followed by a grid search
to fit a mortality model
}

