% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitmethod-combined.R
\docType{data}
\name{combinedFit}
\alias{combinedFit}
\title{fitMethod object for fitting via combination of
optim and grid search}
\format{fitMethod object}
\usage{
combinedFit
}
\description{
fitMethod object for fitting via combination of
optim and grid search
}
\keyword{datasets}

