\name{create.mortalityFits}
\alias{create.mortalityFits}
\title{create.mortalityFits}
\usage{
  create.mortalityFits(fits)
}
\arguments{
  \item{fits}{the list of mortalityFit objects}
}
\value{
  a mortalityFits object
}
\description{
  given a list of mortalityFit objects, create a
  mortalityFits object to represent them. first, check to
  be sure that all of the models were fit to the same
  dataset
}

