% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitmethod-gridSearch.R
\docType{data}
\name{gridSearchFit}
\alias{gridSearchFit}
\title{fitMethod object for fitting via grid search}
\format{fitMethod object}
\usage{
gridSearchFit
}
\description{
fitMethod object for fitting via grid search
}
\keyword{datasets}

