% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classdefns.R
\docType{class}
\name{mortalityData-class}
\alias{mortalityData-class}
\title{mortalityData}
\description{
representation of a dataset that contains
}
\details{
\describe{
    \item{name}{the dataset's name}
    \item{age.offset}{the amount to add to the age column of \code{data} in order to obtain the actual ages. For example, the ages in \code{data} might range from 1 to 20 and represent ages 80 to 99; in that case, \code{age.offset} would be 79}
    \item{age.interval.width}{the width, in years, of each age interval. for now, this is always assumed to be 1. future versions may accommodate different age intervals.}
    \item{data}{a data.frame with at least three columns:
          \describe{
              \item{age}{the age (last birthday)}
              \item{Nx}{the amount of exposure at the given age}
              \item{Dx}{the number of deaths at the given age}
          }
    }
    \item{tags}{a list whose key/value entries can be used to store metadata; useful for grabbing subsets of a big database. For example, this might have \code{list(country=\"belgi\", sex=\"m\")}}
}
}

