% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classdefns.R
\docType{class}
\name{mortalityFits-class}
\alias{compare.fits.mortalityFits}
\alias{mortalityFits-class}
\title{mortalityFits}
\usage{
compare.fits.mortalityFits(mf, ...)
}
\arguments{
\item{mf}{a data frame with the fit results, as produced
by as.data.frame.mortalityFits}

\item{...}{other params, not currently  used}
}
\value{
a list with entries df, which is a dataframe with
       the fit results, and summary
}
\description{
representation of a list of several
fits of mortalityHazard objects to
the same data

given a list of mortalityFit objects, contained in a
mortalityFits object, produce comparisons
between them along a number of dimensions.
NB/TODO: for now, assuming that all were fit
to the same data. eventually, it would be nice to explicitly
check this
}
\details{
\describe{
    \item{name}{the name of this set of fits}
    \item{data}{the mortalityData object that all of these models were fit to}
    \item{fits}{a list whose entries contain the fits of all of the models}
}
}

