% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classdefns.R
\docType{class}
\name{mortalityModel-class}
\alias{mortalityModel-class}
\title{mortalityModel}
\description{
this object represents a likelihood function for
use with mortality data. generally, it will be
the result of combining a likelihood with a hazard;
for example, it could be the binomial model with the gompertz hazard
embedded in it. many of the slots of this object correspond
to those found in \code{\link{mortalityHazard-class}}
}
\details{
\describe{
    \item{name}{the model's name}
    \item{loglik.fn}{the log-likelihood function, which takes as its arguments
       \describe{
          \item{theta}{a vector of parameter values}
          \item{Dx}{vector of deaths by age}
          \item{Nx}{vector of exposure by age}
          \item{ages}{vector of ages that corresponds to the entries of \code{Nx} and \code{Dx}}
      } }
   \item{num.param}{the number of parameters in the model}
   \item{theta.default}{default values for the parameters}
   \item{theta.range}{if not NULL, a range of plausible values that the parameters can assume}
   \item{optim.default}{a list of default parameters to pass into R's \code{optim} function}
   \item{predict.fn}{a function which, given values for the parameters and a dataset of exposures, produces predicted numbers of deaths by age. it takes as its arguments:
        \describe{
          \item{theta}{a vector of parameter values}
          \item{Nx}{vector of exposure by age}
          \item{age}{vector of ages that the \code{Nx} correspond to}
          \item{age.offset}{the offset used for \code{age}. (See \code{\link{mortalityData-class}}.)}
        } }
   \item{theta.start.fn}{if not NULL, a function which takes a mortalityData object and returns very preliminary estimates of the parameters which can then be used in an optimization routine}
   \item{simulate.fn}{if not NULL, a function which takes as its arguments:
        \describe{
          \item{theta}{a vector of parameter values}
          \item{Nx}{vector of exposure by age}
          \item{age}{vector of ages that the \code{Nx} correspond to}
          \item{seed}{if not NULL, the value to seed the pseudorandom number generator with}
          \item{means}{TODO}
        } }
    \item{eval.fn}{if not NULL, a function which produces an evaluation of the model's fit to a given dataset. it takes as its arguments:
        \describe{
          \item{fit.obj}{the mortalityFit object that resulted from fitting the model to a dataset}
          \item{observed}{the mortalityData object with the observed data}
          \item{fittedValues}{the mortalityPrediction object with fitted values, usually resulting from a call to \code{predict.fn}}
        } }
    \item{hazard}{if not NULL, the mortalityHazard object used in making this mortalityModel}
}
}

