% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/classdefns.R
\docType{class}
\name{mortalityPrediction-class}
\alias{mortalityPrediction-class}
\title{mortalityPrediction}
\description{
this object contains a set of fitted values,
generated from a model's prediction
}
\details{
\describe{
    \item{name}{the name of this mortalityPrediction object}
    \item{fitted.Dx}{vector with the estimated number of deaths by age}
    \item{fitted.qx}{vector with the estimated probabilities of death by age}
    \item{theta}{vector of parameter values used for these predictions}
    \item{age}{vector of ages corresponding to the entries of \code{fitted.Dx} and \code{fitted.Dx}}
    \item{age.offset}{the offset used for ages; for example, if \code{age} has values 1 to 20 and \code{age.offset} is 79, then the real ages being studied range from 80 to 99.}
    \item{Nx}{the amount of exposure, for each age, used in producing these predictions}
}
}

