% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mortfit-help.R
\docType{package}
\name{mortfit}
\alias{mortfit}
\alias{mortfit-package}
\title{Fitting parametric hazard models}
\description{
mortfit has a set of tools useful in fitting parametric
hazard models to datasets that have counts of deaths.
}

