% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitmethod-optim.R
\name{optim.fit}
\alias{optim.fit}
\title{optim.fit}
\usage{
optim.fit(model.obj, data, theta.init = NULL, random.start = FALSE,
  verbose = FALSE, auto.retry = TRUE, ...)
}
\arguments{
\item{model.obj}{a mortalityModel object}

\item{data}{a mortalityData object}

\item{theta.init}{if not NULL, starting values to use for theta}

\item{random.start}{if TRUE, and the model.obj has
a non-NULL theta.range, then
draw a random starting value
uniformly from the values specified
by theta.range}

\item{verbose}{if TRUE, print extra info}

\item{auto.retry}{if TRUE, then if convergence isn't reached, automatically
retry with another draw of random starting values}

\item{...}{TODO}
}
\value{
a mortalityFitOptim object
}
\description{
use optim to fit a mortality model
}

