% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitmethod-optim-multiplestarts.R
\name{optim.multiple.fit}
\alias{optim.multiple.fit}
\title{optim.multiple.fit}
\usage{
optim.multiple.fit(model.obj, data, M = 5, verbose = TRUE,
  ignore.folded = FALSE, ...)
}
\arguments{
\item{model.obj}{a mortalityModel object}

\item{data}{a mortalityData object}

\item{M}{the number of random starts (in addition to defaults)}

\item{verbose}{set to TRUE to print extra information while fitting (useful for debugging)}

\item{ignore.folded}{TODO}

\item{...}{other params to pass to optimFit fit method}
}
\value{
a mortalityFitOptimMultiple object
}
\description{
fit optim using the hazard default thetas and
several random starts; pick the best result
}

