% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fitmethod-optim.R
\docType{data}
\name{optimFit}
\alias{optimFit}
\title{fitMethod object for fitting via optim}
\format{fitMethod object}
\usage{
optimFit
}
\description{
fitMethod object for fitting via optim
}
\keyword{datasets}

