% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitmethod-optim-multiplestarts.R
\docType{data}
\name{optimMultipleFit}
\alias{optimMultipleFit}
\title{fitMethod object for fitting via optim using default 
parameters and several random starts, then picking best result}
\format{fitMethod object}
\usage{
optimMultipleFit
}
\description{
fitMethod object for fitting via optim using default 
parameters and several random starts, then picking best result
}
\keyword{datasets}

