% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotmethods.R
\name{plot.mortalityData}
\alias{plot.mortalityData}
\title{method to plot mortality data}
\usage{
plot.mortalityData(x, y = NULL, ...)
}
\arguments{
\item{x}{the mortalityData object to plot}

\item{y}{(not used)}

\item{...}{TODO}
}
\description{
method to plot mortality data
}

