% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/classdefns.R
\docType{methods}
\name{[[,mortalityFits-method}
\alias{[[,mortalityFits-method}
\title{[[ function for mortalityFits object}
\usage{
\S4method{[[}{mortalityFits}(x, i, j, ...)
}
\description{
[[ function for mortalityFits object
}

