% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitmethods.R
\name{summarize.fit}
\alias{summarize.fit}
\title{summarize.fit}
\usage{
summarize.fit(fit.obj)
}
\arguments{
\item{fit.obj}{the mortalityFit object to summarize}
}
\description{
summarize the fit of a model to data,
create a fitSummary object with
the results, and store the fit
summary object in the fit object
}

