% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classdefns.R
\name{as.data.frame.mortalityFit}
\alias{as.data.frame.mortalityFit}
\title{as.data.frame.mortalityFit}
\usage{
as.data.frame.mortalityFit(obj)
}
\arguments{
\item{obj}{the mortalityFit object to convert to a data.frame}
}
\value{
a data.frame containing the info from the
        mortalityFit object
}
\description{
convert a mortalityFit object to a data.frame
this function is intended to be called from
as.data.frame.mortalityFits (ie, from a list of
data frames)
}

