% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotmethods.R
\name{plot.mortalityDataWithFits}
\alias{plot.mortalityDataWithFits}
\title{method to plot a set of mortalityFit objects and the
data they were fitted to (results in a grid of plots)}
\usage{
plot.mortalityDataWithFits(x, y = NULL, Dx = TRUE, ncol = NULL,
  singleplot = FALSE, pdffile = NULL, ...)
}
\arguments{
\item{x}{mortalityFit object}

\item{y}{not used}

\item{Dx}{if TRUE, plot in terms of deaths; otherwise, use central death rates}

\item{ncol}{if not NULL, the number of columns in the array of plots produced}

\item{singleplot}{if TRUE, plot all of the fit objects on the same plot; otherwise,
                  return a list of plots}

\item{pdffile}{the name of a .pdf file to save to (if NULL, plot is displayed to the screen)}

\item{...}{other args, which are passed on to the pdf device}
}
\description{
method to plot a set of mortalityFit objects and the
data they were fitted to (results in a grid of plots)
}

