% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotmethods.R
\name{plot.mortalityDataWithFits}
\alias{plot.mortalityDataWithFits}
\title{method to plot a set of mortalityFit objects and the
data they were fitted to (results in a grid of plots)}
\usage{
plot.mortalityDataWithFits(x, y = NULL, Dx = TRUE, ncol = NULL,
  singleplot = FALSE, pdffile = NULL, ...)
}
\description{
@param x mortalityFit object
 @param y not used
 @param Dx if TRUE, plot in terms of deaths; otherwise, use central death rates
 @param singleplot if TRUE, plot all of the fit objects on the same plot; otherwise,
                   return a list of plots
 @param ncol if not NULL, the number of columns in the array of plots produced
 @param pdffile the name of a .pdf file to save to (if NULL, plot is displayed to the screen)
 @param ... other args, which are passed on to the pdf device
 @export
}

