\docType{methods}
\name{plot}
\alias{plot}
\title{generic plot method for mortfit package}
\arguments{
  \item{x}{req'd as part of the generic defn of plot}

  \item{y}{req'd as part of the generic defn of plot}

  \item{...}{used in some contexts}
}
\description{
  generic plot method for mortfit package
}
\seealso{
  \code{\link{plot.mortalityDataWithFit},
  \link{plot.mortalityData},
  \link{plot.mortalityDataWithHazard},
  \link{plot.mortalityHazard},
  \link{plot.mortalityDataWithFit}}
}

