\name{summarize.fit}
\alias{summarize.fit}
\title{summarize.fit}
\usage{
  summarize.fit(fit.obj)
}
\arguments{
  \item{fit.obj}{the mortalityFit object to summarize}
}
\description{
  summarize the fit of a model to data, create a fitSummary
  object with the results, and store the fit summary object
  in the fit object
}

